refactor(site): split the canvas theme so its skins run under bun test
circ-theme.mjs decoded five PNGs at import, which needs new Image(), so
bun test could never load it and its one test read the file as text. The
palettes move to circ-palette.mjs, every drawing function to
circ-skins.mjs behind makeSkins(assets), which takes the sprites and any
offscreen canvas by injection; circ-theme.mjs stays the entry the islands
import and binds the page's live sprite lookup. The moved bodies are the
same except that a sprite is read through sprite(name).

A recording 2D context drives every skin at three cell sizes, three
signals, two widths, with and without sprites, and pins what it drew as
an op-log golden per kind, with the wire and highlight hooks beside them;
UPDATE_GOLDENS=1 rewrites them. Nothing the canvas draws changes.
